changes col to decimal for test

+++ data/migrations/20190418094937_saferoutes.js
@@ -5,8 +5,8 @@ exports.up = function(knex) {
       accidents.string('TWAY_ID', 128);
       accidents.string('TWAY_ID2', 128);
       accidents.string('COUNTY', 128);
-      accidents.float('LATITUDE', 128);
-      accidents.float('LONGITUD', 128);
+      accidents.decimal('LATITUDE', 128);
+      accidents.decimal('LONGITUD', 128);
       accidents.integer('MONTH', 128);
       accidents.integer('DAY', 128);
       accidents.integer('YEAR', 128);
